Remove whitelist hooks, and modify feathers service, write swagger
related to Giveth/giveth-dapp#2286
